Slippery Marker's Dynamic Videoplayer System [SMDVS] V1.0
went back to the drawing board and decided that this code will remain as a "single purpose" utility, it will play videos in a random order like ads on a billboard (while muted). the videos' addresses are kept in an array which can be modified from the editor to avoid hard coding the addresses.